+++ src/mmds/src/data_store.rs
@@ -1,24 +1,36 @@
 // Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::token::{Error as TokenError, TokenAuthority};
 use crate::MAX_DATA_STORE_SIZE;
 use serde_json::{to_vec, Value};
 use std::fmt;
+use std::fmt::{Display, Formatter};
 
 /// The Mmds is the Microvm Metadata Service represented as an untyped json.
 #[derive(Clone)]
 pub struct Mmds {
     data_store: Value,
+    // None when MMDS V1 is configured, Some for MMDS V2.
+    token_authority: Option<TokenAuthority>,
     is_initialized: bool,
     data_store_limit: usize,
-    version: MmdsVersion,
 }
 
 /// MMDS version.
-#[derive(Clone, Copy)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub enum MmdsVersion {
     V1,
     V2,
+}
+
+impl Display for MmdsVersion {
+    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+        match self {
+            MmdsVersion::V1 => write!(f, "V1"),
+            MmdsVersion::V2 => write!(f, "V2"),
+        }
+    }
 }
 
 /// MMDS possible outputs.
@@ -27,20 +39,26 @@ pub enum OutputFormat {
     Imds,
 }
 
-#[derive(Debug, PartialEq)]
+#[derive(Debug)]
 pub enum Error {
+    AlreadyAtRequestedVersion(MmdsVersion),
     DataStoreLimitExceeded,
     NotFound,
     NotInitialized,
+    TokenAuthority(TokenError),
     UnsupportedValueType,
 }
 
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
+        match &*self {
+            Error::AlreadyAtRequestedVersion(ver) => {
+                write!(f, "MMDS is already using requested version {}", ver)
+            }
             Error::DataStoreLimitExceeded => write!(f, "The MMDS patch request doesn't fit."),
             Error::NotFound => write!(f, "The MMDS resource does not exist."),
             Error::NotInitialized => write!(f, "The MMDS data store is not initialized."),
+            Error::TokenAuthority(err) => write!(f, "Token Authority error: {}", err),
             Error::UnsupportedValueType => write!(
                 f,
                 "Cannot retrieve value. The value has an unsupported type."
@@ -53,9 +71,9 @@ impl Default for Mmds {
     fn default() -> Self {
         Mmds {
             data_store: Value::default(),
+            token_authority: None,
             is_initialized: false,
             data_store_limit: MAX_DATA_STORE_SIZE,
-            version: MmdsVersion::V1,
         }
     }
 }
@@ -72,14 +90,52 @@ impl Mmds {
         }
     }
 
-    /// Set the MMDS version.
-    pub fn set_version(&mut self, version: MmdsVersion) {
-        self.version = version;
+    /// Checks if the provided token has not expired.
+    pub fn is_valid_token(&self, token: &str) -> Result<bool, TokenError> {
+        match &self.token_authority {
+            None => Err(TokenError::NotInitialized),
+            Some(ta) => Ok(ta.is_valid(token)),
+        }
     }
 
-    /// Return the MMDS version.
+    /// Set the MMDS version.
+    pub fn set_version(&mut self, version: MmdsVersion) -> Result<(), Error> {
+        match version {
+            MmdsVersion::V1 => self.set_mmds_v1(),
+            MmdsVersion::V2 => self.set_mmds_v2(),
+        }
+    }
+
+    /// Sets MMDS version back to V1 by removing the token authority entity.
+    fn set_mmds_v1(&mut self) -> Result<(), Error> {
+        match self.token_authority {
+            None => Err(Error::AlreadyAtRequestedVersion(MmdsVersion::V1)),
+            Some(_) => {
+                self.token_authority = None;
+                Ok(())
+            }
+        }
+    }
+
+    /// Sets MMDS version to V2 by creating the token authority entity.
+    fn set_mmds_v2(&mut self) -> Result<(), Error> {
+        match self.token_authority {
+            None => {
+                let token_authority = TokenAuthority::new().map_err(Error::TokenAuthority)?;
+                self.token_authority = Some(token_authority);
+                Ok(())
+            }
+            Some(_) => Err(Error::AlreadyAtRequestedVersion(MmdsVersion::V2)),
+        }
+    }
+
+    /// Return the MMDS version by checking the token authority field.
     pub fn version(&self) -> MmdsVersion {
-        self.version
+        if self.token_authority.is_none() {
+            MmdsVersion::V1
+        } else {
+            MmdsVersion::V2
+        }
     }
 
     pub fn set_data_store_limit(&mut self, data_store_limit: usize) {
@@ -260,12 +316,16 @@ mod tests {
 
         // Test invalid path.
         assert_eq!(
-            mmds.get_value("/invalid_path".to_string(), OutputFormat::Json),
-            Err(Error::NotFound)
+            mmds.get_value("/invalid_path".to_string(), OutputFormat::Json)
+                .unwrap_err()
+                .to_string(),
+            Error::NotFound.to_string()
         );
         assert_eq!(
-            mmds.get_value("/invalid_path".to_string(), OutputFormat::Imds),
-            Err(Error::NotFound)
+            mmds.get_value("/invalid_path".to_string(), OutputFormat::Imds)
+                .unwrap_err()
+                .to_string(),
+            Error::NotFound.to_string()
         );
 
         // Retrieve an object.
@@ -296,8 +356,9 @@ mod tests {
         assert_eq!(
             mmds.get_value("/age".to_string(), OutputFormat::Imds)
                 .err()
-                .unwrap(),
-            Error::UnsupportedValueType
+                .unwrap()
+                .to_string(),
+            Error::UnsupportedValueType.to_string()
         );
 
         // Test path ends with /; Value is a dictionary.
@@ -316,8 +377,9 @@ mod tests {
         assert_eq!(
             mmds.get_value("/phones/".to_string(), OutputFormat::Imds)
                 .err()
-                .unwrap(),
-            Error::UnsupportedValueType
+                .unwrap()
+                .to_string(),
+            Error::UnsupportedValueType.to_string()
         );
 
         // Test path does NOT end with /; Value is a dictionary.
@@ -329,8 +391,9 @@ mod tests {
         assert_eq!(
             mmds.get_value("/phones".to_string(), OutputFormat::Imds)
                 .err()
-                .unwrap(),
-            Error::UnsupportedValueType
+                .unwrap()
+                .to_string(),
+            Error::UnsupportedValueType.to_string()
         );
 
         // Retrieve the first element of an array.
@@ -354,8 +417,9 @@ mod tests {
         assert_eq!(
             mmds.get_value("/member".to_string(), OutputFormat::Imds)
                 .err()
-                .unwrap(),
-            Error::UnsupportedValueType
+                .unwrap()
+                .to_string(),
+            Error::UnsupportedValueType.to_string()
         );
 
         // Retrieve a float.
@@ -367,8 +431,9 @@ mod tests {
         assert_eq!(
             mmds.get_value("/shares_percentage".to_string(), OutputFormat::Imds)
                 .err()
-                .unwrap(),
-            Error::UnsupportedValueType
+                .unwrap()
+                .to_string(),
+            Error::UnsupportedValueType.to_string()
         );
 
         // Retrieve a negative integer.
@@ -380,8 +445,9 @@ mod tests {
         assert_eq!(
             mmds.get_value("/balance".to_string(), OutputFormat::Imds)
                 .err()
-                .unwrap(),
-            Error::UnsupportedValueType
+                .unwrap()
+                .to_string(),
+            Error::UnsupportedValueType.to_string()
         );
     }
 
@@ -439,8 +505,8 @@ mod tests {
         let data = "{\"new_key2\" : \"smth\"}";
         let data_store: Value = serde_json::from_str(&data).unwrap();
         assert_eq!(
-            mmds.patch_data(data_store).unwrap_err(),
-            Error::DataStoreLimitExceeded
+            mmds.patch_data(data_store).unwrap_err().to_string(),
+            Error::DataStoreLimitExceeded.to_string()
         );
         assert!(!mmds.get_data_str().contains("smth"));