+++ CHANGELOG.md
@@ -25,6 +25,8 @@
   `X-metadata-token`, which accepts a string value that provides a session
   token for MMDS requests; and `X-metadata-token-ttl-seconds`, which
   specifies the lifetime of the session token in seconds.
+- `GET` requests to MMDS require a session token to be provided through
+  `X-metadata-token` header when MMDSv2 is configured.
 
 ### Changed